requirements.txt
asgiref==3.8.1
Django==5.1.6
django-allauth==65.4.1
django-cors-headers==4.7.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
phonenumbers==8.13.55
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.3
